% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_year_photoper.R
\name{bio_year_photoper}
\alias{bio_year_photoper}
\title{Delineate biological years defined by a photoperiod based biofix}
\usage{
bio_year_photoper(
  dates,
  latitude,
  longitude,
  retain_photoper = TRUE,
  start_previous_year = FALSE,
  shortest_day = FALSE,
  yday_head,
  bio_year_head
)
}
\arguments{
\item{dates}{Vector of dates in class "Date".}

\item{latitude}{Latitude in decimal degrees WGS84.}

\item{longitude}{Longitude in decimal degrees WGS84.}

\item{retain_photoper}{Should the column of daylengths be returned as part of
the result? Boolean.}

\item{start_previous_year}{Does the biological year start during the previous
calendar year? Boolean.}

\item{shortest_day}{Is the biofix defined by the shortest day of the year?
Boolean.}

\item{yday_head}{Header to be assigned to a biological year day column (e.g.,
salmonid_yday).}

\item{bio_year_head}{Header to be assigned to a biological year column (e.g.,
salmonid_year).}
}
\value{
A dataframe consisting of original date column in addition to
columns: photoperiod (if specified), biological year day, biological year.
}
\description{
This function takes a column of dates and a latitude and longitude to
facets/delineates biological years based on photoperiod conditions that
define a "biofix". For example, the start (biofix) of the Atlantic salmon
(smolt) year occurs on the shortest day of the previous calendar year.
}
\examples{
bio_year_photoper(dates = seq(from = as.Date("1981-01-01"),
                             to = as.Date("1984-01-31"), by = "day"),
                 latitude = 53.932458,
                 longitude = -9.575556,
                 start_previous_year = TRUE,
                 shortest_day = TRUE,
                 yday_head = "salmonid_yday",
                 bio_year_head = "salmonid_year")
}
