% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_year_therm.R
\name{bio_year_therm}
\alias{bio_year_therm}
\title{Delineate biological years defined by a biofix temperature}
\usage{
bio_year_therm(
  mean_temp,
  dates,
  biofix_temp,
  min_no_days_above_biofix,
  increasing_temp,
  yday_head,
  bio_year_head,
  days_since_earliest_biofix_head,
  incomplete_first_year = NULL,
  start_previous_calendar_year,
  ...
)
}
\arguments{
\item{mean_temp}{Vector of daily mean temperatures in degrees Celsius.}

\item{dates}{Vector of dates in class "Date".}

\item{biofix_temp}{Biofix temperature.}

\item{min_no_days_above_biofix}{Minimum number of days for which biofix
temperature must be exceeded}

\item{increasing_temp}{Is temperature rising; i.e., does the biofix occur in
spring? Boolean. Note FALSE implies biofix occurs as temperatures are
falling (i.e., in autumn) - FALSE is experimental.}

\item{yday_head}{Header to be assigned to year day column.}

\item{bio_year_head}{Header to be assigned to biological year column.}

\item{days_since_earliest_biofix_head}{Header to be assigned to a universal
biological year day column, assuming that biofix temperature is not reached
on the same day every year. Biological year days in this column are
relative to the earliest biofix year day recorded in the time series (e.g.,
01 June 1990; year day 152)}

\item{incomplete_first_year}{Specified bio year at start of time series for
which data are incomplete.}

\item{start_previous_calendar_year}{Does the biological year start during the
previous calendar year; e.g., as temperatures are cooling during autumn
(e.g., anadromous brown trout smolts that migrate to sea during spring may
begin downstream movements the previous autumn).}

\item{...}{Additional arguments to nested functions.}
}
\value{
A dataframe of original date and temperatures in addition to columns:
year day, biological year, biological year day, comparative biological year
day relative to earliest biofix in the time series.
}
\description{
This function takes a date and temperature column and creates a
new column that facets/delineates biological years based on thermal
conditions that define a "biofix". For example, the function can be used to
identify the start (biofix) of the European eel biological year, which
might occur when temperatures are rising and have exceeded X degrees
Celsius for a minimum of Y consecutive days.
}
\examples{
\dontrun{
data_water_temp <- data.frame(
  "date" = fishcastr::data_ERA5_1979_2019_Jun_bc$date,
  "water_temp" = fishcastr::air_to_water_model(
    Ta = fishcastr::data_ERA5_1979_2019_Jun_bc$tas,
    Yday = lubridate::yday(data_ERA5_1979_2019_Jun_bc$date),
    A = fishcastr::air_to_water_Feeagh_params_ERA5_bcc$A,
    ac = fishcastr::air_to_water_Feeagh_params_ERA5_bcc$ac,
    b = fishcastr::air_to_water_Feeagh_params_ERA5_bcc$b,
    B = fishcastr::air_to_water_Feeagh_params_ERA5_bcc$B
  )
)

data_seel_forecast_years <-
  fishcastr::bio_year_therm(
    mean_temp = data_water_temp$water_temp,
    dates = data_water_temp$date,
    biofix_temp = 11,
    min_no_days_above_biofix = 10,
    increasing_temp = TRUE,
    yday_head = "eel_yday",
    bio_year_head = "eel_year",
    days_since_earliest_biofix_head = "eel_yday_biofix",
    incomplete_first_year = 1978,
    start_previous_calendar_year = FALSE
  )
}
}
