% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_phenology.R
\name{boxplot_phenology}
\alias{boxplot_phenology}
\title{Generate boxplot illustrating migration phenology anomalies from daily
ecological counts.}
\usage{
boxplot_phenology(
  bio_year,
  counts,
  yday,
  method = "mean",
  species_name,
  season,
  dates,
  unreliable_years = NULL,
  plot = TRUE,
  mean_tercile_dates = NULL,
  average_perc_per_tercile = NULL,
  detrended = TRUE,
  summary_stat_calculation_months = NULL,
  ...
)
}
\arguments{
\item{bio_year}{A vector of biological years concurrent with daily counts.}

\item{counts}{A vector of daily counts.}

\item{yday}{A vector of biologically relevant year days.}

\item{method}{A character string, specifying the method for calculating
anomalies (e.g., "mean","median",or "33_66"). Default "mean". Can also
choose a percentile of interest, (currently: "5th", "25th", "33rd", "66th",
"75th" and "95") for the percentiles of fish migration run. IMPORTANT -
Note that the anomaly colour codings in plots will relate directly to the
chosen method, and will therefore not necessarily represent anomalies for
central measures (mean, medians), particularly where data are not unimodal.}

\item{species_name}{A character string.}

\item{season}{A character string.}

\item{dates}{A vector of class "Date".}

\item{unreliable_years}{A concatenated list of unreliable years (e.g., years
during which floods led to unreliable silver eel counts)}

\item{plot}{Boolean indicating if plot should be returned (e.g., in addition
to returned data). See note for method for explanation of colours.}

\item{mean_tercile_dates}{A list of matrices containing (moving) average
biological year day quantiles (e.g., average year day of 33rd and 66th
percentiles and upper and lower CIs). This argument is only required for
comparing observed and simulated phenology. Default NULL.}

\item{average_perc_per_tercile}{A matrix of average (median) percentages of
counts observed in each tercile with upper and lower bootstrapped CIs. This
argument is only required for comparing observed and simulated phenology.
Default NULL.}

\item{detrended}{Boolean to determine if quantile trends are taken into
account; i.e., if detrended, the biological year day delineating the mean
+- 95\% CIs may vary among years with a linear trend. Default TRUE.}

\item{summary_stat_calculation_months}{The moths for which counts should be
used to calculate migration summary statistics (e.g., MAMJ; c(3:6))}

\item{...}{Additional arguments to nested functions.}
}
\value{
A list. "summary_stats", A 12 X no. years summary table, containing
summary statistics for ecological time series count data; namely: mean,
median, and 5th, 25th, 33rd, 50th, 66th, 75th, 95th percentile biological
year days - this table also contains the assigned anomaly for each year in
the series and plots it colour coded (i.e., blue: early; red: late; yellow:
no different to average); "mean_terciles", the long term (or moving)
average position of the 33rd and 66th percentiles of the selected summary
statistic (e.g., the mean); "earliest_date_yday", the lowest calendar year
day (i.e., not forecast year day) of the time series (e.g., 60 for salmon
smolts, which is 60th day of the year, not 60 days post winter solstice) -
this is not used within the function; "summary_stat_calculation_months",
returns the selected target season in a concatenated list of months (e.g.,
2:7).
}
\description{
This function takes vectors of dates, counts, biological years
and biological year days and illustrates inter-annual variation in
phenological anomalies relative to long term (full series average; mean or
median) or moving average expectations (year days on which mean or median
is expected to occur follow a linear trend). Anomalies are illustrated
using a colour coded boxplot, which illustrates whether each summary
statistic (e.g., mean or median) is assigned to lower, middle or upper
tercile; that is, during a defined central expectation window (e.g., mean
+- 95\% CI, which may shift linearly with time). We expect one third of
means to be recorded in each tercile and we define three anomalies: early,
late, average. The user can specify if anomalies are calculated using the
mean-tercile method (described above), or a variation of this method, the
median-tercile, or based on bulk movements of 33.3\% and 66.6\% percentiles
(as described in fishcastr::tercile_plot_phenology()). Can also choose a
percentile of interest, (e.g., 0.05) for the beginning of a fish migration
run. The purpose of this plot is to illustrate to the user that arbitrary
selection of summary statistics to infer long term trends in phenology can
be misleading in cases where phenology curves are not symmetrical (e.g., in
diadromous fish count time series). Bulk movements may be more useful for
multimodal phenology curves.
}
\examples{
\dontrun{
 data_counts <-
   readRDS(paste0(
    system.file("vignettes", package = "fishcastr"),
    "/vignette_data/data_ssmolt_enviro_1981_2019.rds"
  ))

 data_counts_subset <- data_counts[data_counts$salmonid_year \%in\% c(1981:2018),]

# boxplot and return list
 test <- fishcastr::boxplot_phenology(
     bio_year = data_counts_subset$salmonid_year,
     counts = data_counts_subset$ssmolt,
     yday = data_counts_subset$salmonid_yday,
     dates = data_counts_subset$date,
     method = "mean",
     species_name = "Atlantic salmon (smolt)",
     season = "Spring",
     summary_stat_calculation_months = c(3:7),
     plot = TRUE,
     detrended = FALSE
   )
}
}
