% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_terciles.R
\name{calculate_terciles}
\alias{calculate_terciles}
\title{Identify year days of recorded quantiles for multi-year count time series}
\usage{
calculate_terciles(bio_year, counts, yday, terciles, unreliable_years)
}
\arguments{
\item{bio_year}{A vector of biological years concurrent with daily counts.}

\item{counts}{A vector of daily counts.}

\item{yday}{A vector of biologically relevant year days.}

\item{terciles}{A vector of length = 2, detailing the percentile based
central window of the count season of interest e.g., (0.33, 0.66).}

\item{unreliable_years}{A concatenated list of unreliable years (e.g., years
during which floods allow fish to bypass traps).}
}
\value{
A list of length 2. terciles_species_year; A 3 x no. years matrix of
tercile dates. Each row relates to one year CI95_se_terciles; A 2 x 3
matrix of mean year days for each tercile in addition to 95\% CIs.
}
\description{
This function takes vectors of counts, biological years and
biological year days and identifies the year day on which two or three
percentiles (e.g., 33 and 66) are recorded in multi-year ecological count
time series. In addition, the multi-year average expectation of each
tercile and its 95\%CI is returned.
}
\examples{
\dontrun{
data_counts <-
  readRDS(paste0(
    system.file("vignettes", package = "fishcastr"),
    "/vignette_data/data_ssmolt_enviro_1981_2019.rds"
  ))
data_counts_sub <-
  data_counts[data_counts$salmonid_year \%in\% c(1999:2018), ]
calculate_terciles(
  unreliable_years = c(2000, 2004),
  bio_year = data_counts_sub$salmonid_year,
  counts = data_counts_sub$ssmolt,
  yday = data_counts_sub$salmonid_yday,
  terciles = c(1 / 3, 2 / 3)
)
}
}
