% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_terciles_detrended.R
\name{calculate_terciles_detrended}
\alias{calculate_terciles_detrended}
\title{Identify year days of recorded quantiles for multi-year count time series}
\usage{
calculate_terciles_detrended(
  bio_year,
  counts,
  yday,
  terciles,
  unreliable_years
)
}
\arguments{
\item{bio_year}{A vector of biological years concurrent with daily counts.}

\item{counts}{A vector of daily counts.}

\item{yday}{A vector of biologically relevant year days.}

\item{terciles}{A vector of length = 2, detailing the percentile based
central window of the count season of interest e.g., (0.33, 0.66).}

\item{unreliable_years}{A concatenated list of unreliable years (e.g., years
during which floods led to unreliable silver eel counts).}
}
\value{
A list of length 6: terciles_species_year; A 3 x no. years matrix of
tercile dates. Each row relates to one year CI95_se_terciles.
CI95_se_terciles_list_trend; A list of length equal to number of years in
time series, Each list element contains a 2 x 3 matrix of fitted (mean
expected) terciles (e.g., 33rd and 66th) and each's accompanying 95\%CIs.
lower_perc_lm; A lm object of the lower tercile (e.g., 33.3rd percentile)
lower_perc_mean_95CI; A 3 x no. years matrix of fitted lower tercile year
days in addition to 95\%CIs upper_perc_lm;  A lm object of the upper tercile
(e.g., 66.6th percentile) upper_perc_mean_95CI; A 3 x no. years matrix of
fitted upper tercile year days in addition to 95\% CIs.
}
\description{
This function takes vectors of counts, biological years and
biological year days and identifies the year day on which two quantiles
(e.g., 33 and 66) are recorded in multi-year ecological count time series.
In addition, the multi-year moving average expectation of each tercile and
its 95\% CI is returned in the form of a list. Moving averages are calculated
from OLS lm objects, which are also returned by the function.
}
\examples{
\dontrun{
data_counts <-
  readRDS(paste0(
    system.file("vignettes", package = "fishcastr"),
    "/vignette_data/data_ssmolt_enviro_1981_2019.rds"
  ))
data_counts_sub <-
  data_counts[data_counts$salmonid_year \%in\% c(1999:2018), ]
calculate_terciles_detrended(
  unreliable_years = c(2000, 2004),
  bio_year = data_counts_sub$salmonid_year,
  counts = data_counts_sub$ssmolt,
  yday = data_counts_sub$salmonid_yday,
  terciles = c(1 / 3, 2 / 3)
)
}
}
