% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cauchy_resp.R
\name{Cauchy_resp}
\alias{Cauchy_resp}
\title{Cauchy curve}
\usage{
Cauchy_resp(gamma, x0, x)
}
\arguments{
\item{gamma}{A constant that defines the kurtosis, spread and scale of the
curve.}

\item{x0}{A constant that defines the position of the curve peak.}

\item{x}{A numeric vector.}
}
\value{
y A numeric vector.
}
\description{
This function describes a unimodal curve defined by y values as a function of
x and two constants. For example, y might define a proxy measure for
migration preparedness for marine endurance of diadromous fish migrating into
sea water following morphological transformation (e.g., silvering of European
eels). The Cauchy curve is heavy-tailed and is suitable for response
variables for which outliers occur at smaller and larger values of x than
expected under a Gaussian curve.
}
