% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_grid_to_dataframe.R
\name{convert_grid_to_dataframe}
\alias{convert_grid_to_dataframe}
\title{Convert a grid object to a data.frame}
\usage{
convert_grid_to_dataframe(
  grid_obj,
  output_dir = NULL,
  output_folder_name = "output",
  ...
)
}
\arguments{
\item{grid_obj}{A list of grids (can be single grid).}

\item{output_dir}{A file path}

\item{output_folder_name}{Output parent folder name (to be created within
specified path)}

\item{...}{additional arguments (e.g., header == FALSE to pass to
write.table)}
}
\value{
A data.frame (if single member). A list of data.frames if
multi-member grid.
}
\description{
This function takes a list of grid objects (e.g., multiple meteorological
variables downloaded and interpolated from climate reanalysis or multi-member
seasonal climate forecast) and extracts variable information and concurrent
dates. Optionally export output to a .dat file
}
\examples{
\dontrun{
# single member reanalysis
grid_sub <- lapply(fishcastr::grid_ERA5_1979_2019_Jan_bc,
                   function(x){
                   transformeR::subsetGrid(x,
                                             years = 1996:1998)})
convert_grid_to_dataframe(grid_obj = grid_sub,
                          output_dir = paste0(getwd(),
                                              "/inst/extdata/ERA5_archive_1979_2019"))

# multi-member seasonal forecast
grid_sub_mm <- lapply(fishcastr::grid_SEAS5_1993_2019_2_3_4_5_6_7_8_tas_pr_petH_bc,
                   function(x){
                     transformeR::subsetGrid(x,
                                             years = 1996:1998)})
convert_grid_to_dataframe(grid_obj = grid_sub_mm,
                          output_dir = paste0(getwd(),
                                              "/inst/extdata/SEAS5_archive_1993_2019"))
# single member reanalysis
grid_sub <- lapply(fishcastr::grid_ERA5_1979_2019_Jan_bc,
                   function(x){
                   transformeR::subsetGrid(x,
                                             years = 1996:1998)})
convert_grid_to_dataframe(grid_obj = grid_sub)

# multi-member seasonal forecast
grid_sub_mm <- lapply(fishcastr::grid_SEAS5_1993_2019_2_3_4_5_6_7_8_tas_pr_petH_bc,
                   function(x){
                     transformeR::subsetGrid(x,
                                             years = 1996:1998)})
convert_grid_to_dataframe(grid_obj = grid_sub_mm)
}
}
