% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_multi_member_grid_to_list.R
\name{convert_multi_member_grid_to_list}
\alias{convert_multi_member_grid_to_list}
\title{Convert a climate multi-member grid to a nested list}
\usage{
convert_multi_member_grid_to_list(multi_member_grid)
}
\arguments{
\item{multi_member_grid}{A multi-member grid.}
}
\value{
A nested list of years and multi-variable data.frame members within
years.
}
\description{
This function takes a multi-member grid object (e.g., as
seasonal climate forecast grid) and converts it to a nested list, whereby
each primary list element refers to a year and secondary list elements
relate to members within years. Each member is a multivariable data.frame.
}
\examples{
\dontrun{
 mm_grid <- fishcastr::grid_SEAS5_1993_2019_2_3_4_5_6_7_8_tas_pr_petH_bc
 mm_grid_sub <- lapply(mm_grid,FUN = function(x){
   transformeR::subsetGrid(grid = x,years = 1996:1998)})
 test <- convert_multi_member_grid_to_list(
   multi_member_grid = mm_grid_sub
 )
 }
}
