% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/convert_stratified_census_data_to_model_warmup_list.R
\name{convert_stratified_census_data_to_model_warmup_list}
\alias{convert_stratified_census_data_to_model_warmup_list}
\title{A function to subset stratified fish census data into lists for warm-up of
hydrologic or water temperature models}
\usage{
convert_stratified_census_data_to_model_warmup_list(
  water_temp,
  photoper,
  forecast_years,
  forecast_yday,
  dates,
  no_years_warmup,
  warmup_from_solstice = TRUE,
  delete_op_year = TRUE
)
}
\arguments{
\item{water_temp}{A numeric vector}

\item{photoper}{A numeric vector.}

\item{forecast_years}{A numeric vector.}

\item{forecast_yday}{A numeric vector.}

\item{dates}{A date vector.}

\item{no_years_warmup}{An integer value.}

\item{warmup_from_solstice}{Default TRUE (no other options at this time)}

\item{delete_op_year}{Remove operational year data. Default is TRUE as the
first use of this function in forecasting workflow is to calibrate a model
using historic data.}
}
\value{
A two column data.frame with headers: "date" and "pwdds".
}
\description{
This function takes an input data.frame with at least date and
forecast year columns. This function is applicable to cases where count
data (e.g., migration runs) are stratified into separate years, but
antecedent environmental conditions may extend backwards into the previous
forecast year. For example, hydrologic conditions in spring 1981 are
appropriately modelled using a warm-up period that extends backwards into
1980 or even 1979. Likewise, in relation to European eels, migration runs
are stratified in years that run from ~May to April, but antecedent thermal
conditions (e.g., degree days) might influence the onset and rate of
silvering transition (e.g., a run starting in 1981 may be influenced by
thermal conditions experienced since winter solstice 1980). The current
default here is to start warm-ups from winter solstice; see examples for
years that start on winter solstice (i.e., salmonids in Burrishoole) and
forecast years that start post winter solstice (i.e, eels in Burrishoole).
}
\examples{

# LOAD MET DATA ----
met_data <- fishcastr::grid_ERA5_1979_2019_Jan_bc
met <-
  fishcastr::convert_grid_to_dataframe(grid_obj = met_data)[, -2]
names(met)[which(names(met) == "dates1")] <-
  "date"

# RUN AIR TO WATER TEMP MODEL -----
data_water_temp <- data.frame(
   "date" = met$date,
   "Tw_mod" = fishcastr::air_to_water_model(
    Ta = met$tas,
    Yday = lubridate::yday(met$date),
    A = fishcastr::air_to_water_Feeagh_params_ERA5_bcc$A,
    ac = fishcastr::air_to_water_Feeagh_params_ERA5_bcc$ac,
    b = fishcastr::air_to_water_Feeagh_params_ERA5_bcc$b,
    B = fishcastr::air_to_water_Feeagh_params_ERA5_bcc$B
  )
)

# STRATIFY DATA BY FORECAST YEAR (RETAIN PHOTOPERIOD) ----
data_ssmolt_forecast_years <-
  fishcastr::bio_year_photoper(
    dates = data_water_temp$date,
    latitude = 53.932458,
    longitude = -9.575556,
    retain_photoper = TRUE,
    start_previous_year = TRUE,
    shortest_day = TRUE,
    yday_head = "salmonid_yday",
    bio_year_head = "salmonid_year"
  )

# CALCULATE PHOTOPERIOD WEIGHTED DEGREE DAYS ----
  convert_stratified_census_data_to_model_warmup_list(
    water_temp = data_water_temp$Tw_mod,
    photoper = data_ssmolt_forecast_years$photoper,
    forecast_years = data_ssmolt_forecast_years$salmonid_year,
    forecast_yday = data_ssmolt_forecast_years$salmonid_yday,
    dates = data_ssmolt_forecast_years$date,
    no_years_warmup = 0
  )

}
