% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ewembi_1979_2016.R
\docType{data}
\name{data_ewembi_1979_2016}
\alias{data_ewembi_1979_2016}
\title{EWEMBI pseudo- meteorological observations at Burrishoole, Co. Mayo, Ireland.}
\format{
A data table with six columns: \describe{
\item{date}{a date, Gregorian calendar}
\item{tas_ewembi}{mean air temperature, in degrees Celsius}
\item{pr_ewmebi}{total precipitation, in millimetres}
\item{tasmin_ewmebi}{min air temperature, in degrees Celsius}
\item{tasmax_ewmebi}{max air temperature, in degrees Celsius}
\item{petH_ewembi}{potential evapotranspiration, in millimetres}... }
}
\source{
\url{"https://meteo.unican.es/tds5/catalogs/ewembi/ewembiDatasets.html?dataset=ewembi/ewembi_daily.ncml"}
\url{https://www.isimip.org/gettingstarted/input-data-bias-correction/details/27/}
}
\usage{
data_ewembi_1979_2016
}
\description{
A data table containing daily mean, min and max air temperature at 2m and
daily total precipitation and potential evapotranspiration (calculated using
the Hargreaves-Samani equation) observed at Burrishoole, obtained from the
EWEMBI gridded observation dataset. Dates 1979-01-01 to 2016-12-31. These
data may be used to fill gaps in weather station data.
}
\keyword{datasets}
