% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Feeagh2m_temp.R
\docType{data}
\name{data_Feeagh2m_temp}
\alias{data_Feeagh2m_temp}
\title{Daily mean water temperature, Lough Feeagh, Burrishoole, Co. Mayo, Ireland.}
\format{
A data frame with 5843 rows and two variables: \describe{
\item{date}{date, a date} \item{Water_Temp_2m}{water temperature, in degrees
Celsius} ... }
}
\source{
\url{http://data.marine.ie/geonetwork/srv/eng/catalog.search#/metadata/ie.marine.data:dataset.3757}
}
\usage{
data_Feeagh2m_temp
}
\description{
A dataset containing daily mean water temperature (at 2m depth) recorded at
the Automatic Water Quality monitoring station (AWQMS) since 2004-01-01.
}
\keyword{datasets}
