% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_MR_temp.R
\docType{data}
\name{data_MR_temp}
\alias{data_MR_temp}
\title{Midnight water temperature in the Mill Race, Burrishoole, Co. Mayo, Ireland.}
\format{
A data frame with 18262 rows and two variables: \describe{
\item{date}{date, a date} \item{meanT}{midnight temperature, in degrees
Celsius} ... }
}
\source{
\url{http://data.marine.ie/geonetwork/srv/eng/catalog.search#/metadata/ie.marine.data:dataset.2796}
}
\usage{
data_MR_temp
}
\description{
A dataset containing midnight water temperature recorded at the Mill Race
since 1970-01-01.
}
\keyword{datasets}
