% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_seel.R
\docType{data}
\name{data_seel}
\alias{data_seel}
\title{Daily counts of downstream migrating European (silver) eels recorded at
Burrishoole, Co. Mayo, Ireland.}
\format{
A data frame with 18262 rows and two variables: \describe{
\item{date}{date, a date} \item{seel}{number of silver eels, an integer}
... }
}
\source{
\url{http://data.marine.ie/geonetwork/srv/eng/catalog.search#/metadata/ie.marine.data:dataset.4343}
}
\usage{
data_seel
}
\description{
A dataset containing daily counts of downstream migrating European (silver)
eels recorded at the Mill Race and Salmon Leap fish traps since 1970-01-01.
}
\keyword{datasets}
