% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ssmolt.R
\docType{data}
\name{data_ssmolt}
\alias{data_ssmolt}
\title{Daily counts of downstream migrating Atlantic salmon (smolts) recorded at
Burrishoole, Co. Mayo, Ireland.}
\format{
A data frame with 18262 rows and two variables: \describe{
\item{date}{date, a date} \item{ssmolt}{number of salmon smolts, an
integer} ... }
}
\source{
\url{http://data.marine.ie/geonetwork/srv/eng/catalog.search#/metadata/ie.marine.data:dataset.4343}
}
\usage{
data_ssmolt
}
\description{
A dataset containing daily counts of downstream migrating Atlantic salmon
(smolts) recorded at the Mill Race and Salmon Leap fish traps since
1970-01-01.
}
\keyword{datasets}
