% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_stsmolt.R
\docType{data}
\name{data_stsmolt}
\alias{data_stsmolt}
\title{Daily counts of downstream migrating Anadromous brown trout (smolts) recorded
at Burrishoole, Co. Mayo, Ireland.}
\format{
A data frame with 18262 rows and two variables: \describe{
\item{date}{date, a date} \item{stsmolt}{number of trout smolts, an
integer} ... }
}
\source{
\url{http://data.marine.ie/geonetwork/srv/eng/catalog.search#/metadata/ie.marine.data:dataset.4343}
}
\usage{
data_stsmolt
}
\description{
A dataset containing daily counts of downstream migrating Anadromous brown
trout (smolts) recorded at the Mill Race and Salmon Leap fish traps since
1970-01-01.
}
\keyword{datasets}
