% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_since_biofix.R
\name{days_since_biofix}
\alias{days_since_biofix}
\title{Delineate ordinal biological year days relative to earliest biofix in time
series}
\usage{
days_since_biofix(
  dataset,
  bio_date_day,
  calendar_date,
  bio_year,
  year_day,
  start_previous_calendar_year,
  ...
)
}
\arguments{
\item{dataset}{A dataframe cintaining dates, temperatures, iological year
day, biological year, calendar year day}

\item{bio_date_day}{Header of column in dataset that contains biological year
days.}

\item{calendar_date}{Header of column in dataset that contains calendar
dates.}

\item{bio_year}{Header of column in dataset containing biological years.}

\item{year_day}{Header of column in dataset containing calendar ordinal year
days.}

\item{start_previous_calendar_year}{Does the biological year start during the
previous calendar year (e.g., anadromous brown trout, which migrate in
spring following smoltification which may begin during previous calendar
year)? Boolean.}

\item{...}{Additional arguments to nested functions.}
}
\value{
A vector of ordinal biological year days relative to earliest biofix
in the time series.
}
\description{
This function takes a dataset of dates, temperatures, biological
year days, biological years, and ordinal year days and returns a vector of
biological year days relative to the earliest ordinal year day on which a
thermal biofix was reached in the dataset. Thus, a inter-annually
comparable biological year day is calculated that can be used to compare
the timing of ecological events (e.g, migration phenology).
}
