% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_var.R
\name{delta_var}
\alias{delta_var}
\title{Calculate rate of change "delta" for a variable}
\usage{
delta_var(y, react_time, ...)
}
\arguments{
\item{y}{A vector (e.g., daily water temperature).}

\item{react_time}{Number of days of changing variable perceivable by fish.}

\item{...}{Additional arguments to nested functions.}
}
\value{
A vector of daily delta values.
}
\description{
This function calculates a daily gradient defined by a specified
number of points. The user specifies the number of days for which the
gradient will be calculated (e.g., 10 days). The number of days specified
relates to the assumed perceptive abilities of diadromous fishes. For
example, one might expect eels to migrate in response to decreasing
temperature, but eels might only perceive a decrease in temperature if is
occurs over several days or weeks; i.e., eels might not migrate in response
to a decrease in temperature from one day to the next, but rather respond
to an average rate of decrease in temperature that has occurred over a
specified period.
}
\examples{
delta_var(y = rnorm(n = 365, mean = 10 + 10*sin((1:365*(pi/365.25)) + (1.5*pi*(1/365.25))),
                    sd = 2),
          react_time = 20)
}
