% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_EWEMBI.R
\name{download_EWEMBI}
\alias{download_EWEMBI}
\title{Download EWEMBI pseudo-climate observations for Lough Feeagh, Burrishoole,
Ireland}
\source{
\url{https://dataservices.gfz-potsdam.de/pik/showshort.php?id=escidoc:3928916}
}
\usage{
download_EWEMBI()
}
\value{
A message indicating successful or unsuccessful download.
}
\description{
A function to download and pre-process Potsdam Institute for
Climate Impact Research's EWEMBI dataset, which was produced for the ISIMIP
project. See URL for further details. This function downloads data stored
in multi-year netcdf files and builds a grid object compatible with the
University of Cantabria Met Group's Climate4R tools. The EWEMBI data were
primarily constructed for bias correction of climate predictions in the
ISIMIP project (See https://esg.pik-potsdam.de/search/isimip/ for more
details about ISIMIP and EWEMBI); however, we used them in fishcastr for
infilling gaps in meteorological station air temperature data in
Burrishoole, Co. Mayo, Ireland. The data are accessed via the THREDDS Data
Server (TDS).

@details After running this function, two resulting .rds datasets (one in grid form
and one in tabular form) are stored in the extdata folder in the fishcastr
package directory and contain daily mean, max and min air temperature (at
2m depth), total daily precipitation, and calculated potential
evapotranspiration (by the Hargreaves-Samani method) from 1979-01-01 to
2016-12-31. Moreover, a tabular .dat file will be stored in the
extdata/ISIMIP_EWEMBI/EWEMBI_potsdam folder. The raw EWEMBI data were
published under the CC By 4.0 license.

An alternative method for obtaining these data is via the University of
Cantabria Met Group's User Data Gateway, which requires registration to
gain access credentials. This method is contained in a script in the
data-raw/data_met_obs folder.
}
\examples{
\dontrun{
download_EWEMBI()
}
}
