% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_Feeagh_wlevel.R
\name{download_Feeagh_wlevel}
\alias{download_Feeagh_wlevel}
\title{Download water level for Lough Feeagh, Burrishoole, Ireland and estimate
discharge}
\source{
\url{https://www.epa.ie/hydronet/#32070}
}
\usage{
download_Feeagh_wlevel()
}
\value{
A message indicating successful or unsuccessful download.
}
\description{
A function to download and pre-process water level recordings
for Lough Feeagh, Burrishoole, Co. Mayo, Ireland to produce estimates for
catchment discharge. Execution of this function results in two .rds
datasets stored in the extdata folder in the package directory:
data_Feeagh_discharge and data_Feeagh_discharge_corr. Each contains rating
curve derived daily mean discharge calculated from the daily mean water
level (delineated at 00:00) that have been aggregated from ~15 minutely
records from the EPA's gauge records 1976-03-03 to present; however, one of
the files data_Feeagh_discharge_corr, contains a step adjustment to the raw
EPA water level data, whereby 0.072 m has been added to all values
pre-1996-08-04. The position of this step adjustment and its magnitude is
calculated using the strucchange package. This function also fills any data
gaps of 2 days or fewer with simple linear interpolation.

The rating curve data and model fit is contained in this package.

Note the raw dataset downloaded when this function is run is continually
updated by the EPA, but most recent records have not been checked for
quality; hence, this function retains only checked values. On 2021-04-26,
most recent checked value was 2020-10-07. To retain reproducibility of the
resultant datasets (and the step change), we subset each dataset at this
date.
}
\examples{
\dontrun{
download_Feeagh_wlevel()
}
}
