% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_fish_data.R
\name{download_fish_data}
\alias{download_fish_data}
\title{Download fish count data}
\source{
\url{https://data.marine.ie/geonetwork/srv/eng/catalog.search#/metadata/ie.marine.data:dataset.4343}
}
\usage{
download_fish_data()
}
\value{
A message indicating successful or unsuccessful download.
}
\description{
A function to download daily counts of downstream migrating
diadromous fishes recorded at the Mill Race and Salmon Leap traps in
Burrishoole, Co. Mayo, Ireland since 1970-01-01. The dataset contains
counts of Atlantic salmon smolts, anadromous brown trout smolts (Sea trout
smolts) and European silver eels.
}
\examples{
\dontrun{
download_fish_data()
}
}
