% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_Met_Eireann_station_data.R
\name{download_Met_Eireann_station_data}
\alias{download_Met_Eireann_station_data}
\title{Download Met Éireann's manual and automatic station data for Furnace,
Burrishoole.}
\source{
\url{https://www.met.ie/climate/available-data/historical-data}
}
\usage{
download_Met_Eireann_station_data()
}
\value{
A message indicating successful or unsuccessful download.
}
\description{
A function to download meteorological observations (manual and
automatic) from two stations near the south shore of Lough Feeagh,
Burrishoole, Co. Mayo, Ireland. The manual station data range from
1959-11-01 to present and the auto station data range from 2005-02-22 to
present. This function subsets both the raw manual station data and the
auto station data to 2019-12-31 and calculates mean air temperature from
max and min daily values. Resultant .rds datasets are stored in the extdata
folder of the fishcastr package.
}
\examples{
\dontrun{
download_Met_Eireann_station_data()
}
}
