% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_MR_temperature.R
\name{download_MR_temperature}
\alias{download_MR_temperature}
\title{Download midnight water temperature data for the the Mill Race, Burrishoole,
Ireland}
\source{
\url{https://data.marine.ie/geonetwork/srv/eng/catalog.search#/metadata/ie.marine.data:dataset.2796}
}
\usage{
download_MR_temperature()
}
\value{
A message indicating successful or unsuccessful download.
}
\description{
A function to download water temperature recordings for the Mill
Race, Burrishoole, Co. Mayo, Ireland. The raw data records span 1960-01-01 to
2017-12-31.
}
\examples{
\dontrun{
download_MR_temperature()
}
}
