% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earliest_day.R
\name{earliest_day}
\alias{earliest_day}
\title{Find the year day of the earliest biofix in the time series}
\usage{
earliest_day(
  dataset,
  calendar_date,
  bio_date_day,
  start_previous_calendar_year,
  ...
)
}
\arguments{
\item{dataset}{A dataframe containing dates, temperatures, biological year
day, biological year, calendar year day}

\item{calendar_date}{Header of column in dataset that contains calendar
dates.}

\item{bio_date_day}{Header of column in dataset that contains biological year
days.}

\item{start_previous_calendar_year}{Does the biological year start during the
previous calendar year (e.g., anadromous brown trout, which migrate in
spring following smoltification which may begin during previous calendar
year)? Boolean.}

\item{...}{Additional arguments to nested functions.}
}
\value{
A single year day for the earliest calendar year day on which a
biofix occured.
}
\description{
This function takes a dataset of dates, temperatures, biological
year days, biological years, and ordinal year days and returns the earliest
ordinal year day on which a thermal biofix was reached.
}
