% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_multi_member_grid_to_csv.R
\name{export_multi_member_grid_to_csv}
\alias{export_multi_member_grid_to_csv}
\title{Export a climate multi-member grid to a folder of csvs}
\usage{
export_multi_member_grid_to_csv(
  multi_member_grid,
  name_study,
  dataset_name,
  dir.data = NULL
)
}
\arguments{
\item{multi_member_grid}{A Date.}

\item{name_study}{Character string to be used as folder name.}

\item{dataset_name}{Character string used to label sub folders (e.g., SEAS5).}

\item{dir.data}{Directory path for exported data.}
}
\value{
A nested list of years and members within years.
}
\description{
This function takes a multi-member grid object (e.g., as
seasonal climate forecast) and converts it to a set of folders containing
multivariable forecasts (one per member).
}
\examples{
\dontrun{
mm_grid <- fishcastr::grid_SEAS5_1993_2019_2_3_4_5_6_7_8_tas_pr_petH_bc
export_multi_member_grid_to_csv(
  multi_member_grid = mm_grid,
  dataset_name = SEAS5,
  name_study = "Burrishoole",
  dir.data = system.file("vignettes", package = "fishcastr"),
  members = 1:25
)
}
}
