% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fish_data.R
\name{import_fish_data}
\alias{import_fish_data}
\title{Import fish count data}
\source{
\url{https://data.marine.ie/geonetwork/srv/eng/catalog.search#/metadata/ie.marine.data:dataset.4343}
}
\usage{
import_fish_data(species)
}
\arguments{
\item{species}{A character string choice of salmon smolt "ssmolt", sea trout
smolt "stsmolt" and silver eel "seel".}
}
\value{
A two column data.frame "date" and chosen species label.
}
\description{
A function to import daily counts of downstream migrating
diadromous fishes recorded at the Mill Race and Salmon Leap traps in
Burrishoole, Co. Mayo, Ireland since 1970-01-01. The dataset contains
counts of Atlantic salmon smolts, anadromous brown trout smolts (Sea trout
smolts) and European silver eels. IMPORTANT - this function can only be run
after fishcastr::download_fish_data() has been run.
}
\examples{
\dontrun{
data_ssmolt <- import_fish_data(species = "ssmolt")
}
}
