% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{longestNAstrech}
\alias{longestNAstrech}
\title{Measure length of longest stretch of NAs within a vector}
\usage{
longestNAstrech(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
An integer.
}
\description{
This function takes vectors and finds the length of the longest stretch of
NAs
}
\examples{
x <- rnorm(n = 100, mean = 20, sd = 5)
x[c(10:20,40:50,70:95)] <- NA
longestNAstrech(x) # 100
}
