% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lunar_light_exposure.R
\name{lunar_light_exposure}
\alias{lunar_light_exposure}
\title{Generate proxy measure of moonlight exposure.}
\usage{
lunar_light_exposure(date_of_interest, lat, lon, log_result = TRUE)
}
\arguments{
\item{date_of_interest}{A Date.}

\item{lat}{A numeric.}

\item{lon}{A numeric.}

\item{log_result}{A boolean. Default TRUE.}
}
\value{
A numeric of approximate relative lunar exposure (range 0 to 1
depending on lat argument).
}
\description{
This function takes a date and calculates an approximate measure of moonlight
exposure depending on geographic location. The function extracts the duration
of the night (sunrise to sunset) in minutes and calculates the percentage of
the night during which the moon is above the horizon. The function also
extracts the sunlit fraction of the moon at solar midnight (nadir) and the
moon's peak altitude/angle above the horizon during the night in radians. The
approximate measure of exposure is calculated as ln((sin(peak_moon_altitude)
\itemize{
\item prop_night_moon_above_horizon * illuminated_fraction_nadir^4) + 0.00001) +
min(exposure). This function requires the suncalc package. This proxy does
not account for opposition surge, but the power of 4 is intended to increase
the relative difference between full moon exposure and other phases. However, NOTE:
the log transformation is used to increase resolution of small changes in
lunar light exposure at low exposure values. This can be removed by setting
the log_result argument to FALSE.
}
}
\examples{
lunar_light_exposure(date_of_interest = seq(from = as.Date("1970-01-01"),
                                           to = as.Date("1970-06-30"),
                                           by = "day"),
                    lat = 53.932458,
                    lon = -9.575556)
}
