% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max_bio_year_date.R
\name{min_max_bio_year_date}
\alias{min_max_bio_year_date}
\title{Extract start and end dates of biological/forecast years}
\usage{
min_max_bio_year_date(dates, bio_years, biofix_day)
}
\arguments{
\item{dates}{A vector of Dates.}

\item{bio_years}{A vector of biological years corresponding to dates
argument.}

\item{biofix_day}{A vector containing biological year days corresponding to
dates argument.}
}
\value{
A table containing the start and end date of all biological years
delineated in the input data.
}
\description{
This function extracts the start and end dates of biological
years for use in summary tables. For example, to summarise European eel
"silvering" years.
}
\examples{
\dontrun{

# ----------------------------------------------------------- #
# IMPORT RAW AIR TEMP AND AIR TO WATER TEMP MODEL -----
# ----------------------------------------------------------- #
data_air_temp <-
  fishcastr::convert_grid_to_dataframe(grid_obj = fishcastr::grid_ERA5_1979_2019_Jun_bc)[,-2]
names(data_air_temp) <-
  c("date", "tas_era5_bcc", "pr_era5_bcc", "petH_era5_bcc")
air_to_water_params <-
  fishcastr::air_to_water_Feeagh_params_ERA5_bcc
data_water_temp <- data.frame(
  "date" = data_air_temp$date,
  "water_temp" = fishcastr::air_to_water_model(
    Ta = data_air_temp$tas_era5_bcc,
    Yday = lubridate::yday(data_air_temp$date),
    A = air_to_water_params$A,
    ac = air_to_water_params$ac,
    b = air_to_water_params$b,
    B = air_to_water_params$B
  )
)

# ----------------------------------------------------------- #
# DEFINING FORECAST "THERMAL" YEARS, YEAR DAYS AND DAYS SINCE COMMON BIOFIX
----
# ----------------------------------------------------------- #
eel_bio_year <-
  fishcastr::bio_year_therm(
    mean_temp = data_water_temp$water_temp,
    dates = data_water_temp$date,
    biofix_temp = 11,
    min_no_days_above_biofix = 10,
    increasing_temp = TRUE,
    yday_head = "eel_yday",
    bio_year_head = "eel_year",
    days_since_earliest_biofix_head = "eel_yday_biofix",
    incomplete_first_year = 1978,
    start_previous_calendar_year = FALSE
  )

# ----------------------------------------------------------- #
# SAVE EEL BIOYEAR DATA FOR SUPPLEMENTARY TABLE ----
# ----------------------------------------------------------- #
# Create vignette sub directory ----
tab_eel <-
  knitr::kable(
    fishcastr::min_max_bio_year_date(
      dates = eel_bio_year$date,
      bio_years = eel_bio_year$eel_year,
      biofix_day = eel_bio_year$eel_yday_biofix
    )
  )

kableExtra::kable_styling(kable_input = tab_eel, position = "center")

                 }
}
