% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photper_calc.R
\name{photper_calc}
\alias{photper_calc}
\title{Calculate photoperiod from latitude and longitude}
\usage{
photper_calc(dates, latitude, longitude, ...)
}
\arguments{
\item{dates}{Vector of dates in class "Date".}

\item{latitude}{Latitude in decimal degrees WGS84}

\item{longitude}{Longitude in decimal degrees WGS84}

\item{...}{additional arguments to suncalc::getSunlightTimes}
}
\value{
A vector of daylengths.
}
\description{
This function takes a date vector and latitude and longitude and
calculates photoperiod (daylength) from sunrise and sunset times using the
suncalc package.
}
\examples{
photper_calc(dates = seq(from = as.Date("1981-01-01"),
                        to = as.Date("1981-12-31"), by = "day"),
            latitude = 53.932458,
            longitude = -9.575556)
}
