% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physio_model_evaluation_exp_mod_Gauss.R
\name{physio_model_evaluation_exp_mod_Gauss}
\alias{physio_model_evaluation_exp_mod_Gauss}
\title{Fit an exponentially modified Gaussian response curve to count data}
\usage{
physio_model_evaluation_exp_mod_Gauss(
  data_list,
  species_count_lab,
  x_variable_lab,
  error_distribution_name,
  species_name,
  file_path_plot,
  file_path_resid_fitted_plot,
  file_path_dispersion_zeroinf_acf_plot,
  file_path_coef_profile_plot,
  start_values,
  lower_lims,
  ylims,
  xlims,
  ...
)
}
\arguments{
\item{data_list}{A list of data frames of ecological counts and environmental
variables.}

\item{species_count_lab}{Column header for y variable.}

\item{x_variable_lab}{Column header for x variable.}

\item{error_distribution_name}{One of: "negative_binomial",
"generalised_Poisson", "quasi_Poisson".}

\item{species_name}{A character string label for plot legend.}

\item{file_path_plot}{Character string file path for main figure.}

\item{file_path_resid_fitted_plot}{Character string file path for fitted vs
residual figure.}

\item{file_path_dispersion_zeroinf_acf_plot}{Character string file path for
dispersion and acf diagnostic figures.}

\item{file_path_coef_profile_plot}{Character string file path for bbmle::mle2
model coefficient profile plots}

\item{start_values}{A list of starting values supplied to L-BFGS-B method.}

\item{lower_lims}{A list of lower limit values supplied to L-BFGS-B method.}

\item{ylims}{Concatenated values.}

\item{xlims}{Concatenated values.}

\item{...}{Other arguments to nested functions.}
}
\value{
A bbmle::mle2 model trained on all years of data_list pooled together
(i.e., 5th data_list element)
}
\description{
This function produces a series of plots that can be used to
evaluate the fit of an exponentially modified Gaussian response curve to
unimodal count data. Specifically, plots are intended for checking the
consistency of curve fit (i.e., inter-decadal variability in a proxy for
attainment of migration preparedness) among multiple decades (e.g., fitting
curves for pooled count data from windows 1990 - 2000, 2011 - 2010, 20011 -
2020). Residual plots use the functionality of the DHARMa package (Hartig
2019). The intended implementation of the exponentially modified Gaussian
curve as a proxy for physiological preparedness for migration of diadromous
fishes assumes that physiological preparedness for sea entry increases with
degree days that reflect an increase in physiological condition for marine
endurance (e.g., gill Na+/K+-ATPase activity in salmonids; Sykes et al.,
2009; Zydlewski et al., 2005). See /data-raw/model_physio_expmG_salmon.R
for usage example.
}
\references{
Zydlewski, G. B., Haro, A., & McCormick, S. D. (2005). Evidence for
cumulative temperature as an initiating and terminating factor in downstream
migratory behaviour of Atlantic salmon (Salmo salar) smolts. Canadian Journal
of Fisheries and Aquatic Sciences, 62(1), 68–78.
https://doi.org/10.1139/f04-179
Sykes, G. E., Johnson, C. J., & Shrimpton, J. M. (2009). Temperature and Flow
Effects on Migration Timing of Chinook Salmon Smolts. Transactions of the
American Fisheries Society, 138(6), 1252–1265.
https://doi.org/10.1577/T08-180.1
Hartig, F. (2019). DHARMa: Residual Diagnostics for Hierarchical (Multi-Level
/ Mixed) Regression Models (R package version 0.2.2).
https://CRAN.R-project.org/package=DHARMa
}
