% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_air_to_water_validation.R
\name{plot_air_to_water_validation}
\alias{plot_air_to_water_validation}
\title{Air to water temperature model validation plot.}
\usage{
plot_air_to_water_validation(Tw, Ta, Yday, dates, mod_air_to_water_mle)
}
\arguments{
\item{Tw}{A numeric vector of water temperatures.}

\item{Ta}{A numeric vector of air temperatures.}

\item{Yday}{A numeric vector of year days (derivable from lubridate::yday).}

\item{dates}{A vector of corresponding dates.}

\item{mod_air_to_water_mle}{A fitted bbmle::mle2 object that models the
relationship between air temperature and lake surface water temperature
(load .rds model found in vignettes).}
}
\value{
A matrix of validation figures and a list of summary statistics.
}
\description{
This function returns a variety of validation plots and
validation statistics for the air two water temperature model that builds
on the three-parameter equation defined by A. Ducharne (2008). Ducharne's
equation describes the statistical relationship between lagged rolling
average air temperature and water temperature in streams (whereby lag
period is a constant in days). We extended Ducharne's equation to a
four-parameter model to account for stratification in lakes, whereby the
lag period between air temperature and water temperature shortens during
the summer owing to the shallower mixed layer depth during stratification.
These validation plots are intended for use on out-of-sample air and water
temperature data. See Piccolroaz et al., (2013) for further discussion of
lake thermal dynamics in relation to meteorological forcing, and see and
Calderó-Pascual et al., (2020) for description of Lough Feeagh,
Burrishoole, Ireland example. A bias adjustment multiplication factor that
varies with year day was included to reduce residual hysteresis present in
the basic lagged air temperature model.
}
\references{
Calderó-Pascual, M., de Eyto, E., Jennings, E., Dillane, M., Andersen, M. R.,
Kelly, S., Wilson, H. L., & McCarthy, V. (2020). Effects of Consecutive
Extreme Weather Events on a Temperate Dystrophic Lake: A Detailed Insight
into Physical, Chemical and Biological Responses. Water, 12(5), 1411.
https://doi.org/10.3390/w12051411

Ducharne, A. (2008). Importance of stream temperature to climate change
impact on water quality. Hydrology and Earth System Sciences, 12(3), 797–810.

Piccolroaz, S., Toffolon, M., & Majone, B. (2013). A simple lumped model to
convert air temperature into surface water temperature in lakes. Hydrology
and Earth System Sciences, 17(8), 3323–3338.
https://doi.org/10.5194/hess-17-3323-2013
}
