% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ensemble_data.R
\name{plot_ensemble_data}
\alias{plot_ensemble_data}
\title{Plot time series of ensemble model chain predictions with observations
overlay}
\usage{
plot_ensemble_data(
  seasonal_ensemble,
  pseudo_obs_ensemble,
  var_interest,
  obs_count_name,
  bio_yday,
  bio_year,
  colour_theme,
  ensemble_name,
  target_season,
  lead_month,
  export_to_file = FALSE,
  ensemble.lty = "l",
  plot_climatology = TRUE,
  ylabel,
  export_file_folder,
  multi_panel_plot = FALSE,
  target_season_only = FALSE,
  fill_obs_counts = FALSE
)
}
\arguments{
\item{seasonal_ensemble}{A seasonal forecast list object returned by
fishcastr::compile_seasonal_forecast_glmm.}

\item{pseudo_obs_ensemble}{A seasonal forecast list object returned by
fishcastr::compile_seasonal_forecast_glmm.}

\item{var_interest}{A character string (e.g., "forecast_sim_counts").}

\item{obs_count_name}{A character string (e.g., "ssmolt").}

\item{bio_yday}{A character string (e.g., "salmonid_yday").}

\item{bio_year}{A character string (e.g., "salmonid_year").}

\item{colour_theme}{A character string (e.g., "green3").}

\item{ensemble_name}{A character string (must be "ERA5" or "SEAS5", while
experimental).}

\item{target_season}{An integer concatenated (e.g., 3:7 for March to July).}

\item{lead_month}{An integer (e.g., 1 for January).}

\item{export_to_file}{Boolean. Export all member predictions in addition to
multi-panel plot?}

\item{ensemble.lty}{A character (e.g., "s" for daily counts, "l" for
continuous variables).}

\item{plot_climatology}{Experimental.}

\item{ylabel}{A character string (common y variable name e.g., "fish
counts").}

\item{export_file_folder}{A file path character string.}

\item{multi_panel_plot}{Boolean. plot or not.}

\item{target_season_only}{Boolean. TRUE Crops plot to show only target season
predictions.}

\item{fill_obs_counts}{Boolean. Fills step style polygons for fish counts
grey.}
}
\value{
A data.frame containing all member predictions for all re-forecast
years is exported to a csv. A multi-panel plot is exported to a png.
}
\description{
This function is experimental. The function takes ensemble
seasonal forecast predictions/simulations (e.g., 25 member predictions for
daily fish counts; these are contained in the argument which is the
returned result of fishcastr::compile_seasonal_forecast_glmm). The
multi-panel plots returned by this figure are experimental, but are
intended to inform model checking. The function also returns a table
containing predictions for each member of the forecast for all forecast
years. The plots returned show 95\% percentile prediction intervals and
ensemble medians for the variable of interest.
}
\examples{
\dontrun{
# ERA5 salmon (note code in the salmon workflow vignette must be run to obtain
# this rds object)
seasonal_forecast_list_dfs_ssmolt_daily_stoc_era5 <-
  readRDS(file = paste0(
    system.file("vignettes"),
    "/vignette_data\\\\SF_salmon_ideal_loo_cv_23456.rds"
  ))

# add 2019 counts to forecast list object
fishcastr::download_fish_data()
data_ssmolt <- fishcastr::import_fish_data(species = "ssmolt")

seasonal_forecast_list_dfs_ssmolt_daily_stoc_era5[[39]] <-
  lapply(seasonal_forecast_list_dfs_ssmolt_daily_stoc_era5[[39]], function(x) {
    NA_dates_2019 <- x$date[which(is.na(x$ssmolt))]
    x$ssmolt[x$date \%in\% NA_dates_2019] <-
      data_ssmolt$ssmolt[data_ssmolt$date \%in\% NA_dates_2019]
    return(x)
  })

fish_counts_ERA5 <-
  plot_ensemble_data(
    seasonal_ensemble = seasonal_forecast_list_dfs_ssmolt_daily_stoc_era5,
    pseudo_obs_ensemble = seasonal_forecast_list_dfs_ssmolt_daily_stoc_era5,
    var_interest = "forecast_sim_counts",
    obs_count_name = "ssmolt",
    bio_yday = "salmonid_yday",
    bio_year = "salmonid_year",
    colour_theme = "green3",
    ensemble_name = "ERA5",
    target_season = 3:7,
    lead_month = 1,
    export_to_file = TRUE,
    ensemble.lty = "s",
    plot_climatology = FALSE,
    ylabel = expression("Fish count"),
    export_file_folder = paste0(
      system.file("extdata", package = "fishcastr"),
      "/eel_wflow_figs/"
    ),
    multi_panel_plot = TRUE,
    target_season_only = TRUE,
    fill_obs_counts = TRUE
  )
}
}
