% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential_ev_HS.R
\name{potential_ev_HS}
\alias{potential_ev_HS}
\title{Calculate potential evapotranspiration}
\usage{
potential_ev_HS(tasmin, tasmax, lats, refdates)
}
\arguments{
\item{tasmin}{a vector of min air temperatures}

\item{tasmax}{a vector of max air temperatures}

\item{lats}{Latitude in decimal degrees WGS84}

\item{refdates}{Vector of dates in class "Date".}
}
\value{
A vector of potential evapotranspiration estimates.
}
\description{
This function takes vectors of dates, maximum air temperature
and minimum air temperature and latitude and calculates potential
evapotranspiration by the Hargreaves-Samani equation implemented in
\emph{drought4R} package (Bedia & Iturbide, 2019).
}
\examples{
# simulate min and max temps
min_temp <- rnorm(1:365,
                 mean = 12+8*sin(1.5*pi + seq(length = 365,
                                              from = 0,
                                              to = 2*pi)),
                 sd = 3)
max_temp <- min_temp + 2
potential_ev_HS(refdates = seq(from = as.Date("1981-01-01"),
                               to = as.Date("1981-12-31"), by = "day"),
                tasmin = min_temp,
                tasmax = max_temp,
                lats = 53.932458)
}
\references{
Allen, R. G., & Food and Agriculture Organization of the United Nations
(Eds.). (1998). Crop evapotranspiration: Guidelines for computing crop water
requirements. Food and Agriculture Organization of the United Nations.
Bedia, J., & Iturbide, M. (2019). Drought4R: A climate4R package for drought
assessment. https://github.com/SantanderMetGroup/transformeR/wiki
}
