% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ucount_ci.R
\name{predict_ucount_ci}
\alias{predict_ucount_ci}
\title{Generate simulation based confidence intervals from glmmTMB::glmmTMB model}
\usage{
predict_ucount_ci(model, interval, dataset)
}
\arguments{
\item{model}{A glmmTMB::glmmTMB object.}

\item{interval}{A concatenated list; e.g., c(0.025, 0.975). These percentiles
are extracted using quantiles().}

\item{dataset}{A data.frame with names matching those in formula of object.}
}
\value{
A list of four vectors: pred.ucount.psim, single simulation from
mvrnorm; mean.count, the effective "conditional" (on fixed effects, but not
random effects) mean of 999 mvrnorm simulations; ci.ucount_lwr, the lower
interval percentile (e.g., 0.025); ci.ucount_upr, the upper interval
percentile (e.g., 0.975) of 999 mvrnorm simulations.
}
\description{
This function generates predictions (conditional (on fixed
effect) means) and associated uncertainty intervals from a fitted
glmmTMB::glmmTMB model object. See Brooks et al. (2017) Appendix B and/or
Bolker (2008) for details on population prediction intervals generated
using mvrnorm. Predications are "unconditional", which means that the
variance in random effect parameter estimate is ignored, while the variance
in fixed effects parameters is taken into account.
}
\examples{
\dontrun{
mTrain_genpois <- readRDS(paste0(system.file("vignettes", package = "fishcastr"),
               "/vignette_data/mTrain_genpois_salmon_2019.rds"))
newdata <- mTrain_genpois$frame
newdata <- newdata[newdata$salmonid_year \%in\% c(2012:2013),]
result <- predict_ucount_ci(model = mTrain_genpois,
                            interval = c(0.025,0.975),
                            dataset = newdata)
                            }
}
\references{
Brooks, M. E., Kristensen, K., van Benthem, K. J., Magnusson, A., Berg, C.
W., Nielsen, A., Skaug, H. J., Maechler, M., & Bolker, B. M. (2017). Modeling
Zero-Inflated Count Data With glmmTMB. https://doi.org/10.1101/132753

Brooks, M. (2017). Appendix B: Salamander Example Comparing GLMMs,
Zero-Inflated GLMMs, and Hurdle Models. 11.

Bolker, B. M. (2008). 7.5 Estimating confidence limits of functions of
parameters. In Ecological models and data in R (p. 332). Princeton Univ.
Press.
}
