% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgeline_plot_phenology.R
\name{ridgeline_plot_phenology}
\alias{ridgeline_plot_phenology}
\title{Plot phenology anomalies for multi-year dataset in "ridge line"/"joyplot"
style}
\usage{
ridgeline_plot_phenology(
  bio_year,
  counts,
  yday,
  method = "mean",
  species_name,
  season,
  dates,
  unreliable_years = NULL,
  plot = TRUE,
  mean_tercile_dates = NULL,
  average_perc_per_tercile = NULL,
  detrended = TRUE,
  plot_005_995_perc = TRUE,
  xlims_plot = NULL,
  ...
)
}
\arguments{
\item{bio_year}{A vector of biological years concurrent with daily counts.}

\item{counts}{A vector of daily counts.}

\item{yday}{A vector of biologically relevant year days.}

\item{method}{A character string, specifying the method for calculating
anomalies (i.e., "mean","median",or "33_66"). Default "mean".}

\item{species_name}{A character string.}

\item{season}{A character string.}

\item{dates}{A vector of class "Date".}

\item{unreliable_years}{A concatenated list of unreliable years (e.g., years
during which floods led to unreliable silver eel counts)}

\item{plot}{Boolean indicating if plot should be returned (e.g., in addition
to returned data).}

\item{mean_tercile_dates}{A list of matrices containing (moving) average
biological year day quantiles (e.g., average year day of 33rd and 66th
percentiles and upper and lower CIs). This argument is only required for
comparing observed and simulated phenology. Default NULL.}

\item{average_perc_per_tercile}{A matrix of average (median) percentages of
counts observed in each tercile with upper and lower bootstrapped CIs. This
argument is only required for comparing observed and simulated phenology.
Default NULL.}

\item{detrended}{Boolean to determine if quantile trends are taken into
account; i.e., if detrended, the biological year day delineating the mean
+- 95\% CIs may vary among years with a linear trend. Default TRUE.}

\item{plot_005_995_perc}{Boolean to determine if all data or 0.5th to 99.5th
percentile of non-zero counts are plotted. Default is TRUE. If FALSE, plots
all data.}

\item{xlims_plot}{A concatenated vector defining x axis limits in terms of
count percentiles.}

\item{...}{Additional arguments to nested functions.}
}
\value{
A 9 X no. years summary table, containing summary statistics for
ecological time series count data; namely: mean, median, and 5th, 33rd,
50th, 66th, 95th percentile biological year days.
}
\description{
This function is intended for use with multi-decade ecological
count time series to facilitate illustration of interannual variation in
phenology. The function takes vectors of dates, counts, biological years
and biological year days and illustrates inter-annual variation in
phenological anomalies relative to long term (full series average; mean or
median) or moving average expectations (year days on which mean or median
is expected to occur follow a linear trend). Anomalies are illustrated
using a colour coded joyplot/ridge line plot, which illustrates whether
each summary statistic (i.e., mean or median) is assigned to lower, middle
or upper tercile; that is, during a defined central expectation window
(e.g., mean +- 95\% CI, which may shift linearly with time). We expect one
third of means to be recorded in each tercile and we define three
anomalies: early, late, average. The user can specify if anomalies are
calculated using the mean-tercile method (described above), or a variation
of this method, the median-tercile, or based on bulk movements of 33.3\% and
66.6\% percentiles (as described in tercile_plot_phenology()). The purpose
of this plot is to illustrate to the user that arbitrary selection of
summary statistics to infer long term trends in phenology can be misleading
in cases where phenology curves are not symmetrical (e.g., in diadromous
fish count time series).
}
\examples{
\dontrun{
data_counts <- readRDS(paste0(
    system.file("vignettes", package = "fishcastr"),
    "/vignette_data/data_ssmolt_enviro_1981_2019.rds"
  ))
data_counts_subset <- data_counts[data_counts$salmonid_year \%in\% c(1981:2018),]
test <- ridgeline_plot_phenology(bio_year = data_counts_subset$salmonid_year,
                         counts = data_counts_subset$ssmolt,
                         yday = data_counts_subset$salmonid_yday,
                         dates = data_counts_subset$date,
                         method = "mean",
                         species_name = "Atlantic salmon (smolt)",
                         season = "Spring",
                         plot = TRUE,
                         detrended = TRUE)
}
}
