% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_rand_state.R
\name{set_rand_state}
\alias{set_rand_state}
\title{Restore the random state of global environment}
\usage{
set_rand_state(state)
}
\arguments{
\item{state}{The random seed of the global environment}
}
\value{
The random seed of the global environment
}
\description{
This function restores a previously extracted random state 'seed' of the
global environment.(code from
https://www.r-bloggers.com/2019/08/local-randomness-in-r/).
}
