% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_glm.R
\name{simulate_glm}
\alias{simulate_glm}
\title{Generate simulated values from glmmTMB::glmmTMB model}
\usage{
simulate_glm(
  object,
  nsim = 1,
  seed = NULL,
  X,
  pred_type,
  zinf = FALSE,
  PIs = FALSE,
  lapply_seed = 123,
  parallel = FALSE,
  use_random_seed = FALSE,
  ...
)
}
\arguments{
\item{object}{A glmmTMB::glmmTMB object.}

\item{nsim}{Number of simulations to generate from glmmTMB object.}

\item{seed}{A random seed.}

\item{X}{A data.frame with names matching those in formula of object.}

\item{pred_type}{A character string "predict" or "simulate" to generate
conditional mean, or random deviates, respectively.}

\item{zinf}{Boolean. Default FALSE, but function checks and alters this
automatically in the case of glmmTMB objects.}

\item{PIs}{Boolean. Default FALSE. If "predict", generates mvrnorm based
population prediction intervals (see B Bolker's Ecological Models and Data
with R.); i.e., uncertainty in conditional mean owing to parameter
uncertainty. If "simulate, generates bootstrapped 95\% prediction intervals,
whereby 999 random deviates are generated from a point estimate conditional
mean and 0.025 and 0.975 percentiles are extracted using quantiles().}

\item{lapply_seed}{A random seed used for parallel processing. Default 123.}

\item{parallel}{Boolean. Default FALSE. Use parallel processing?}

\item{use_random_seed}{Boolean. Default FALSE. Experimental.}

\item{...}{Arguments to nested functions.}
}
\value{
If "pred_type = predict", a list containing two elements. The element
"fitted_y" is a list of four vectors ("pred.ucount.sim", a single simulated
mean based on a single mvrnorm sample; "mean.count", the mean calculated
from of 999 mvrnorm simulations (i.e., the conditional (on fixed effects
only) mean of the model; "ci.ucount_lwr", the 0.025 percentile of 999
mvrnorm simulations; "ci.ucount_upr", the 0.975 percentile of 999 mvrnorm
simulations. The element "fitted_y_old" is generated for
checking/comparison and is the predicted (conditional mean) generated by
the "base::predict()" function - this may be removed at a later date. If
"pred_type = simulate", a single vector of simulations is produced (i.e., a
single vector of random deviates based on the model's specified error
distribution; e.g., glmmTMB::nbinom1). This simulation does not take into
account parameter uncertainties, rather, all values are generated assuming
a point estimated mean derived from the "base::predict" function and a
random deviate generator (e.g., gamlss.dist::rNBII). If "PIs = TRUE" AND
"pred_type = simulate", the function returns a list containing three
vectors (95\% bootstrapped percentile confidence intervals derived from 999
random deviates ("low_PI" and "upper_PI") and a bootstrapped "median". If
"PIs = TRUE" AND "pred_type = predict", no additional data are returned.
}
\description{
This function generates simulations from a fitted
glmmTMB::glmmTMB model object when supplied with "newdata". (adapted from
Ben Bolker's answer to a question on stackoverflow under creative commons
(CC-BY-SA 4.0; https://creativecommons.org/licenses/by-sa/4.0/):
https://stackoverflow.com/questions/55230145/passing-new-data-to-simulate-glmmtmb))
}
\examples{
\dontrun{
mTrain_genpois <- readRDS(paste0(system.file("vignettes", package = "fishcastr"),
                                "/vignette_data/mTrain_genpois_salmon_2019.rds"))
newdata <- mTrain_genpois$frame
newdata <- newdata[newdata$salmonid_year \%in\% c(2012:2013),]
result <- simulate_glm(object = mTrain_genpois,
                             X = newdata,
                             pred_type = "predict")
                             }
}
