% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sNBII.R
\name{sNBII}
\alias{sNBII}
\title{Generate summary statistics from a negative binomial type II (quasi-Poisson)
distribution}
\usage{
sNBII(sigma, mu)
}
\arguments{
\item{sigma}{A numeric, dispersion parameter.}

\item{mu}{A vector.}
}
\value{
A vector of random deviates.
}
\description{
This function generates summary statistics when supplied with a vector of
conditional means and dispersion parameter. Implemented for compatibility
with bbmle::mle2 (Bolker & R Core Team 2017), which requires "s" functions
associated with "d" functions. The corresponding discrete density (or
probability mass) function, dNBII, of the negative binomial type II is
contained in package gamlss.dist.
}
\examples{
sNBII(mu = fishcastr::exp_mod_Gaussian_resp(x = 1:365,
                                            c = 10,
                                            mu_exmg = 100,
                                            sigma_exmg = 15,
                                            lamb = 25),
      sigma = 10)
}
\references{
Bolker, B. & R Core Team. (2017). bbmle: Tools for General Maximum Likelihood
Estimation (R package version 1.0.20).
https://CRAN.R-project.org/package=bbmle
}
