% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tercile_plot_phenology.R
\name{tercile_plot_phenology}
\alias{tercile_plot_phenology}
\title{Generate tercile plot illustrating seasonal migration phenology anomalies
from daily ecological counts.}
\usage{
tercile_plot_phenology(
  bio_year,
  counts,
  yday,
  terciles,
  species_name,
  dates,
  unreliable_years = NULL,
  return_data,
  plot = TRUE,
  definitions,
  mean_tercile_dates = NULL,
  average_perc_per_tercile = NULL,
  detrended = TRUE,
  trend_plot = FALSE,
  bulk_pheno_three_anoms = FALSE,
  ...
)
}
\arguments{
\item{bio_year}{A vector of biological years concurrent with daily counts.}

\item{counts}{A vector of daily counts.}

\item{yday}{A vector of biologically relevant year days}

\item{terciles}{A vector of length = 2, detailing the percentile based
central window of the count season of interest e.g., (0.33, 0.66)}

\item{species_name}{A character string, of options ("Atlantic salmon
(smolt)", "Sea trout (smolt)", "European eel (silver)").}

\item{dates}{A vector of class "Date".}

\item{unreliable_years}{A concatenated list of unreliable years (e.g., years
during which floods led to unreliable silver eel counts)}

\item{return_data}{Boolean indicating if function should return the data used
to plot the tercile plot.}

\item{plot}{Boolean indicating if plot should be returned (e.g., in addition
to returned data).}

\item{definitions}{Boolean indicating if interpretive (stand alone) legend
should be returned instead of data based plot.}

\item{mean_tercile_dates}{A list of matrices containing (moving) average
biological year day quantiles (e.g., average year day of 33rd and 66th
percentiles and upper and lower CIs). This argument is only required for
comparing observed and simulated phenology. Default NULL.}

\item{average_perc_per_tercile}{A matrix of average (median) percentages of
counts observed in each tercile with upper and lower bootstrapped CIs. This
argument is only required for comparing observed and simulated phenology.
Default NULL.}

\item{detrended}{Boolean to determine if quanile trends are taken into
account; i.e., if detrended, the biological year day delineating each
tercile may vary among years with a linear trend.}

\item{trend_plot}{Boolean to determine if seasonal trends are plotted. If
TRUE, only trends on positions of quantiles are plotted.}

\item{bulk_pheno_three_anoms}{Only define three anomlalies (late, average and
early). Experimental. Default NULL.}

\item{...}{Additional arguments to nested functions.}
}
\value{
If plot == TRUE, a tercile plot; if trend_plot == TRUE, a trend plot
and; if return_data == TRUE, a multielement list of various matrices/lists
of matrices containing: average_percent_per_tercile; A 3 x 3 matrix of
expected percentage of counts per tercile and 95\% bootstrapped CIs (can
also be given as argument (see above). mean_terciles; A list of 3 x 2
matrices (one per year) detailing the expectated (e.g. moving average mean)
year days of each tercile and 95\% CIs - note these means and CIs are
plotted by trend_plot == TRUE if OLS linear trends are significant.
tercile_data; A 5 x n years matrix showing percentages of counts per
tercile per year in addition to total counts per year. average, extended,
early, late, condensed; 5 separate concatenated lists of each anomaly type,
whereby element values are years during which anomaly occurred.
mean_terciles_raw; A 3 X n years matrix containing year days on which
(e.g., 33.3 and 66.6) quantiles occur and the year during which they occur.
earliest_date; earliest calendar date on which a biofix occurred during the
time series. earliest_date_yday; earliest year day on which a biofix
occurred during the time series.
}
\description{
This function takes vectors of dates, counts, biological years
and biological year days and illustrates inter-annual variation in
phenological anomalies relative to long term (full series consistent
tercile dates) or moving average expectations (year days on which terciles
occur follow a linear trend). Anomalies are illustrated using a tercile
plot, which illustrates percentages of counts recorded during and either
side of a central window of expectation; that is, during a defined central
window (which may shift linearly with time), we expect one third of counts
to be recorded plus minus uncertainty - deviations from expected
percentages of counts recorded during and either side of the central window
define five anomalies: early, late, bimodal, condensed and average. This
function returns a multi-element list object, and three types of plots.
}
\examples{
\dontrun{
data_counts <-
  readRDS(paste0(
    system.file("vignettes", package = "fishcastr"),
    "/vignette_data/data_ssmolt_enviro_1981_2019.rds"
  ))

 data_counts_subset <- data_counts[data_counts$salmonid_year \%in\% c(1981:2018),]
dev.new()
 phenology_anomalies <- tercile_plot_phenology(
   bio_year = data_counts_subset$salmonid_year,
   counts = data_counts_subset$ssmolt,
   yday = data_counts_subset$salmonid_yday,
   dates = data_counts_subset$date,
   terciles = c(1/3, 2/3),
   species_name = "Atlantic salmon (smolt)",
   return_data = TRUE,
   plot = TRUE,
   definitions = FALSE,
   detrended = FALSE,
   trend_plot = FALSE,
)
}
}
