% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tercileBrewerColorRamp.R
\name{tercileBrewerColorRamp}
\alias{tercileBrewerColorRamp}
\title{Define colour ramp for tercile plot}
\usage{
tercileBrewerColorRamp(ncolors)
}
\arguments{
\item{ncolors}{Number of colours.}
}
\value{
A colour ramp.
}
\description{
This function takes a list of colours from RColorBrewer pallette
(e.g., Reds, Greys and Blues) and creates a colour ramp required for a
tercile (image and image.plot based) plot.
}
