% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tercileColor.R
\name{tercileColor}
\alias{tercileColor}
\title{Define colour palette for tercile plot}
\usage{
tercileColor()
}
\value{
A colour ramp.
}
\description{
This function returns a list of colours compatible with
RColorBrewer pallette (e.g., Reds, Greys and Blues)
}
