% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tercileColorRamp.R
\name{tercileColorRamp}
\alias{tercileColorRamp}
\title{Define colour ramp for tercile plot definition plot.}
\usage{
tercileColorRamp(ncolors)
}
\arguments{
\item{ncolors}{Number of colours.}
}
\value{
A colour ramp.
}
\description{
This function takes a list of colours from RColorBrewer pallette
(e.g., Reds, Greys and Blues) and creates a colour ramp required for a
tercile defintion (image and image.plot based) plot.
}
