% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tercileColorRampCustom.R
\name{tercileColorRampCustom}
\alias{tercileColorRampCustom}
\title{Define colour ramp for tercile plot}
\usage{
tercileColorRampCustom(ncolors, custom_cols)
}
\arguments{
\item{ncolors}{Number of colours.}

\item{custom_cols}{A list of three colours in tercile order (low, middle,
high); e.g., c("blue","grey","red").}
}
\value{
A colour ramp.
}
\description{
This function takes a list of three colours (e.g., red, yellow
and blue) and creates three colour ramps from white to the specified
colours required for a tercile (image and image.plot based) plot.
}
