% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tercileColors.R
\name{tercileColors}
\alias{tercileColors}
\title{Define colour palette for tercile plot}
\usage{
tercileColors()
}
\value{
A colour ramp.
}
\description{
This function returns a list of colours compatible with
RColorBrewer pallette (e.g., Reds, Greys and Blues)
}
