% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terciles_plot_data_detrended.R
\name{terciles_plot_data_detrended}
\alias{terciles_plot_data_detrended}
\title{Summarise percentage of counts recorded in each tercile in multi-year time
series, (allowing percentiles to exhibit a linear trend).}
\usage{
terciles_plot_data_detrended(
  counts,
  bio_year,
  yday,
  terciles,
  plot,
  unreliable_years,
  mean_tercile_dates = NULL,
  ...
)
}
\arguments{
\item{counts}{A vector of daily counts.}

\item{bio_year}{A vector of biological years concurrent with daily counts.}

\item{yday}{A vector of biologically relevant year days.}

\item{terciles}{A vector of length = 2, detailing the percentile based
central window of the count season of interest e.g., (0.33, 0.66).}

\item{plot}{Boolean. Should plot of raw counts be returned?}

\item{unreliable_years}{A concatenated list of unreliable years (e.g., years
during which floods led to unreliable silver eel counts).}

\item{mean_tercile_dates}{A vector of biologically relevant year days.}

\item{...}{Additional arguments to be passed to nested functions.}
}
\value{
A list of length 5 ("year", first tercile currently labelled "early";
middle tercile labelled "average"; third tercile labelled "late"; "n",
number of counted individuals per year) that can be combined with
do.call(cbind, result): each list element is of length no. years in time
series.
}
\description{
This function takes vectors of counts, biological years,
biological year days and the timeseries averaged year days on which two
quantiles (e.g., 33 and 66) occur in multi-year ecological count time
series and returnes the percentage of counts that were recorded in each
tercile.
}
\examples{
\dontrun{
data_counts <-
  readRDS(paste0(
    system.file("vignettes", package = "fishcastr"),
    "/vignette_data/data_ssmolt_enviro_1981_2019.rds"
  ))

data_counts_sub <-
  data_counts[data_counts$salmonid_year \%in\% c(1999:2018), ]

# identify mean year days of each tercile
mean_terciles <-
  lapply(
    calculate_terciles_detrended(
      bio_year = data_counts_sub$salmonid_year,
      counts = data_counts_sub$ssmolt,
      yday = data_counts_sub$salmonid_yday,
      terciles = c(1/3, 2/3)
    )[["CI95_se_terciles_list_trend"]],
    round,
    0
  )

result <- terciles_plot_data_detrended(unreliable_years = c(2000,2004,2008),
                                       bio_year = data_counts_sub$salmonid_year,
                                       counts = data_counts_sub$ssmolt,
                                       yday = data_counts_sub$salmonid_yday,
                                       terciles = c(1/3,2/3),
                                       plot = FALSE,
                                       mean_tercile_dates = mean_terciles)
do.call(cbind,result)
}
}
