% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermal_year_days.R
\name{thermal_year_days}
\alias{thermal_year_days}
\title{Define ordinal biological year days based on a vector of temperature based
biological years}
\usage{
thermal_year_days(thermal_year)
}
\arguments{
\item{thermal_year}{Vector of thermally defined biological years.}
}
\value{
A vector of biological year days.
}
\description{
This function takes a vector of thermally defined biological
years and creates a new vector of biological year days.
}
