% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rand_state.R
\name{get_rand_state}
\alias{get_rand_state}
\title{Extract random state of global environment}
\usage{
get_rand_state()
}
\value{
The random seed of the global environment
}
\description{
This function extracts the current random state 'seed' of the
global environment and is used to store the random state before a
simulation before restoring it with set_rand_state after the simulation.
Code attributed to Evgeni Chasnovski and accessed at
https://www.r-bloggers.com/2019/08/local-randomness-in-r/. Code was
originally posted by Evgeni Chasnovski at
http://www.questionflow.org/2019/08/13/local-randomness-in-r/ under
creative commons  (CC-BY-SA 4.0;
https://creativecommons.org/licenses/by-sa/4.0/).
}
