% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_rand_state.R
\name{set_rand_state}
\alias{set_rand_state}
\title{Restore the random state of global environment}
\usage{
set_rand_state(state)
}
\arguments{
\item{state}{The random seed of the global environment}
}
\value{
The random seed of the global environment
}
\description{
This function restores a previously extracted random state
'seed' of the global environment.  Code attributed to Evgeni Chasnovski and
accessed at https://www.r-bloggers.com/2019/08/local-randomness-in-r/. Code
was originally posted by Evgeni Chasnovski at
http://www.questionflow.org/2019/08/13/local-randomness-in-r/ under
creative commons  (CC-BY-SA 4.0;
https://creativecommons.org/licenses/by-sa/4.0/).
}
